// frontend/src/app/model/zaduzenaKnjiga.ts
export class ZaduzenaKnjiga {
    idKnjige: number;
    username: string;
    datumUzimanja: Date;
    datumVracanja: Date;
    brojDana: number;
    slikaKnjige: string;
    nazivKnjige: string;
    autoriKnjige: string;
    rok: number;
    datumUzimanjaString: string;
    datumVracanjaString: string;
    produzeno: boolean;
}